Added blogs with user authentication

+++ src/blogs/blogs.service.ts
@@ -1,9 +1,14 @@
-import { Injectable } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { CreateBlogDto } from './dto/create-blog.dto';
-// import { UpdateUserDto } from './dto/update-user.dto';
+import { UpdateBlogDto } from './dto/update-blog.dto';
 import { Blog } from './entities/blog.entity';
+import { JwtService } from '@nestjs/jwt';
 // import { User } from 'src/user/entities/user.entity';
 
 @Injectable()
@@ -14,13 +19,18 @@ export class BlogsService {
    */
   constructor(
     @InjectRepository(Blog) private readonly blogRepository: Repository<Blog>,
+    private readonly jwtService: JwtService,
   ) {}
 
-  async createBlog(createBlogDto: CreateBlogDto): Promise<Blog> {
+  async createBlog(createBlogDto: CreateBlogDto, token: string): Promise<Blog> {
+    const decodedToken = await this.decodeToken(token);
+    if (!decodedToken.sub) {
+      throw new UnauthorizedException('Invalid token');
+    }
     const blog: Blog = new Blog();
     blog.title = createBlogDto.title;
     blog.description = createBlogDto.description;
-    // blog.createdBy = await this.blogRepository.findOne(userID);
+    blog.createdBy = decodedToken.sub;
     return this.blogRepository.save(blog);
   }
 
@@ -29,20 +39,74 @@ export class BlogsService {
   }
 
   viewBlog(id: number): Promise<Blog> {
-    return this.blogRepository.findOneBy({ id });
+    return this.blogRepository.findOne({ where: { id } });
   }
 
-  // updateBlog(id: number, updateBlogDto: UpdateBlogDto): Promise<Blog> {
-  //   const blog: Blog = new Blog();
-  //   blog.title = updateBlogDto.title;
-  //   blog.description = updateBlogDto.description;
-  //   blog.id = id;
-  //   return this.blogRepository.save(blog);
-  // }
+  async updateBlog(
+    id: number,
+    updateBlogDto: UpdateBlogDto,
+    token: string,
+  ): Promise<Blog> {
+    const decodedToken = await this.decodeToken(token);
+    if (!decodedToken.sub) {
+      throw new UnauthorizedException('Invalid token');
+    }
 
-  removeBlog(id: number): Promise<{ affected?: number }> {
+    // const blog = await this.blogRepository.findOne({ where: { id } });
+    const blog = await this.blogRepository.findOne({
+      where: { id },
+      relations: ['createdBy'], // Specify the relation to include
+    });
+
+    if (!blog) {
+      throw new NotFoundException('Blog not found');
+    }
+
+    if (blog.createdBy.id !== decodedToken.sub) {
+      throw new UnauthorizedException(
+        'You are not authorized to update this blog',
+      );
+    }
+
+    // Perform partial update based on fields provided in the DTO
+    if (updateBlogDto.title !== undefined) {
+      blog.title = updateBlogDto.title;
+    }
+    if (updateBlogDto.description !== undefined) {
+      blog.description = updateBlogDto.description;
+    }
+
+    return this.blogRepository.save(blog);
+  }
+
+  async removeBlog(id: number, token: string): Promise<{ affected?: number }> {
+    const decodedToken = await this.decodeToken(token);
+    if (!decodedToken.sub) {
+      throw new UnauthorizedException('Invalid token');
+    }
+
+    const blog = await this.blogRepository.findOne({
+      where: { id },
+      relations: ['createdBy'], // Specify the relation to include
+    });
+
+    if (!blog) {
+      throw new NotFoundException('Blog not found');
+    }
+
+    if (blog.createdBy.id !== decodedToken.sub) {
+      throw new UnauthorizedException(
+        'You are not authorized to delete this blog',
+      );
+    }
+
     return this.blogRepository.delete(id);
   }
+
+  // async findAllBlogsByUser(userId: number): Promise<Blog[]> {
+  //   const allBlogs = await this.blogRepository.find();
+  //   return allBlogs.filter((blog) => blog.createdBy.id === userId);
+  // }
 
   async findOne(title: string): Promise<Blog | undefined> {
     return this.blogRepository.findOne({ where: { title: title } });
@@ -51,5 +115,14 @@ export class BlogsService {
   async getAllBlogs(): Promise<Blog[]> {
     // const user: UserEntity = await UserEntity.findOne({where: {id: 2}, relations: ['books']});
     return this.blogRepository.find();
+  }
+
+  private async decodeToken(token: string): Promise<any> {
+    try {
+      const decoded = await this.jwtService.verifyAsync(token);
+      return decoded;
+    } catch (error) {
+      throw new UnauthorizedException('Invalid token');
+    }
   }
 }
